+++ src/main/java/dao/TransportCompanyDAO.java
@@ -1,5 +1,6 @@
 package dao;
 
+import entity.Client;
 import entity.Employee;
 import entity.TransportCompany;
 import jakarta.persistence.EntityManager;
@@ -24,27 +25,40 @@ public class TransportCompanyDAO implements Dao<TransportCompany> {
 
     @Override
     public void update(long id, TransportCompany transportCompany) {
-        entityManager.getTransaction().begin();
+        if (isValid(id)) {
+            entityManager.getTransaction().begin();
 
-        TransportCompany retrievedTransportCompany = entityManager.find(entity.TransportCompany.class, id);
-        retrievedTransportCompany.setName(transportCompany.getName());
-        retrievedTransportCompany.setAddress(transportCompany.getAddress());
-        retrievedTransportCompany.setTotalIncome(transportCompany.getTotalIncome());
-        retrievedTransportCompany.setVehicles(transportCompany.getVehicles());
-        retrievedTransportCompany.setEmployees(transportCompany.getEmployees());
-        retrievedTransportCompany.setTransportations(transportCompany.getTransportations());
-        retrievedTransportCompany.setClients(transportCompany.getClients());
+            TransportCompany retrievedTransportCompany = entityManager.find(entity.TransportCompany.class, id);
+            retrievedTransportCompany.setName(transportCompany.getName());
+            retrievedTransportCompany.setAddress(transportCompany.getAddress());
+            retrievedTransportCompany.setTotalIncome(transportCompany.getTotalIncome());
+            retrievedTransportCompany.setVehicles(transportCompany.getVehicles());
+            retrievedTransportCompany.setEmployees(transportCompany.getEmployees());
+            retrievedTransportCompany.setTransportations(transportCompany.getTransportations());
+            retrievedTransportCompany.setClients(transportCompany.getClients());
 
-        entityManager.getTransaction().commit();
+            entityManager.getTransaction().commit();
+        }
     }
 
     @Override
     public void delete(long id) {
-        entityManager.getTransaction().begin();
+        if (isValid(id)) {
+            entityManager.getTransaction().begin();
 
-        TransportCompany retrievedTransportCompany = entityManager.find(TransportCompany.class, id);
+            TransportCompany retrievedTransportCompany = entityManager.find(TransportCompany.class, id);
 
-        entityManager.remove(retrievedTransportCompany);
-        entityManager.getTransaction().commit();
+            entityManager.remove(retrievedTransportCompany);
+            entityManager.getTransaction().commit();
+        }
+    }
+
+    @Override
+    public boolean isValid(long id) {
+        String idQuery = String.format("SELECT * FROM client WHERE id = %s", id);
+
+        int resultCount = entityManager.createNativeQuery(idQuery, Client.class).getResultList().size();
+
+        return resultCount > 0;
     }
 }
